+++ src/i18n/locales/index.ts
@@ -10,7 +10,6 @@ export { default as theme } from './en/theme.json';
 export { default as profile } from './en/profile.json';
 export { default as entityConfig } from './en/entityConfig.json';
 export { default as modals } from './en/modals.json';
-export { default as navigation } from './en/navigation.json';
 export { default as database } from './en/database.json';
 export { default as config } from './en/config.json';
 export { default as development } from './en/development.json';